+++ digitransit-component/packages/digitransit-component-datetimepicker/src/helpers/utils.js
@@ -1,3 +1,5 @@
+import moment from 'moment-timezone';
+
 /**
  * Handle typing time and adding necessary :
  *
@@ -5,7 +7,7 @@
  *
  * @return {string}
  */
-function parseTypedTime(typedValue) {
+export function parseTypedTime(typedValue) {
   let actualValue = typedValue;
   if (
     actualValue.length === 3 &&
@@ -21,4 +23,90 @@ function parseTypedTime(typedValue) {
   return actualValue;
 }
 
-export default { parseTypedTime };
+const validateClock = (hours, minutes) => {
+  const hoursValid = !Number.isNaN(hours) && hours >= 0 && hours <= 23;
+  const minutesLen = minutes.length;
+  const minutesValid =
+    minutesLen === 2
+      ? !Number.isNaN(minutes) && Number(minutes) >= 0 && Number(minutes <= 59)
+      : !Number.isNaN(minutes) && Number(minutes) >= 0 && Number(minutes <= 5);
+  return hoursValid && minutesValid;
+};
+
+export function validateInput(inputValue) {
+  if (inputValue.length <= 2) {
+    // Too many options, don't  validate
+    return false;
+  }
+  if (inputValue.length === 3) {
+    let hours;
+    let minutes;
+    if (inputValue.includes(':')) {
+      [hours, minutes] = inputValue.split(':');
+    } else if (inputValue.startsWith('0')) {
+      hours = inputValue.substring(0, inputValue.length - 1);
+      minutes = inputValue.substring(inputValue.length - 1 || 0);
+      if (Number(minutes) > 5) {
+        return true;
+      }
+    } else {
+      // This is how basically moment handles string formatting.
+      // If the first letter of the string is 1, then rest are minutes.
+      // if the first letter is 2, then if second letter is 0,1,2 or 3, then second letter is hour, else rest are minutes
+      // else, first letter is hour, rest of them are minutes.
+      const values = inputValue.split('');
+      if (Number(values[0]) === 1) {
+        [hours, minutes] = [values[0].concat(values[1]), values[2]];
+      } else if (Number(values[0]) === 2) {
+        if (Number(values[1] <= 3)) {
+          [hours, minutes] = [values[0].concat(values[1]), values[2]];
+        } else {
+          [hours, minutes] = [values[0], values[1].concat(values[2])];
+        }
+      } else {
+        [hours, minutes] = [values[0], values[1].concat(values[2])];
+      }
+      return !validateClock(hours, minutes);
+    }
+    return !validateClock(hours, minutes);
+  }
+  if (inputValue.length === 5 || inputValue.length === 4) {
+    const values = inputValue.split(':');
+    const hours = values[0];
+    const minutes = values[1];
+    if (
+      inputValue.startsWith('0') &&
+      minutes.length === 1 &&
+      Number(minutes) > 5
+    ) {
+      return true;
+    }
+
+    return !validateClock(hours, minutes);
+  }
+  return false;
+}
+
+export const getTs = (inputValue, currentTimestamp) => {
+  const trimmed = inputValue.trim();
+  if (trimmed.match(/^[0-9]{1,2}(\.|:)[0-9]{2}$/) !== null) {
+    const splitter = trimmed.includes('.') ? '.' : ':';
+    const values = trimmed.split(splitter);
+    const hours = Number(values[0]);
+    const hoursValid = !Number.isNaN(hours) && hours >= 0 && hours <= 23;
+    const minutes = Number(values[1]);
+    const minutesValid =
+      !Number.isNaN(minutes) && minutes >= 0 && minutes <= 59;
+    if (!minutesValid || !hoursValid) {
+      return null;
+    }
+    const newStamp = moment(currentTimestamp)
+      .hours(hours)
+      .minutes(minutes)
+      .valueOf();
+    return newStamp;
+  }
+  return null;
+};
+
+export default { parseTypedTime, validateInput, getTs };
